refactor: test를 위한 구현체를 테스트 패키지로 이동
