+++ src/test/java/model/LineTest.java
@@ -1,7 +1,6 @@
 package model;
 
 import java.util.List;
-import model.strategy.NothingBuildStrategy;
 import model.strategy.RandomBuildStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
